demos/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
python-pptx>=1.0.2
openai>=1.50.0
requests>=2.32.0
python-multipart>=0.0.9
pydantic>=2.9.0
azure-identity>=1.17.0
azure-cosmos>=4.7.0
azure-storage-blob>=12.20.0
python-dotenv>=1.0.1
numpy>=2.0.0
Pillow>=10.0.0
pdf2image>=1.16.0
pywin32>=306;sys_platform=="win32"
websockets>=12.0

# Microsoft Agent Framework (pinned — SDK in preview)
agent-framework-azure-ai
agent-framework-core
azure-ai-agentserver-agentframework
azure-ai-agentserver-core
httpx>=0.27.0
pypdf>=4.0.0
python-docx>=1.1.0
readability-lxml>=0.8.1
azure-search-documents>=11.5.0
